Started dong manim challenges
Completed and "easy" Roulette Epicycloid.
Primary challenge was how to position the dot because manim does not appear to like moving along a curve and rotating objects. Realized you could simulate the circles movement with the movement of the dot, and tied that to a function based on the K of the epicycloid.